third_party/blink/web_tests/external/wpt/focus/iframe-contentwindow-focus-with-same-as-top-intermediate-frame-expected.txt

This is a testharness.js-based test.
[FAIL] Check result
  assert_equals: Check log expected "outeronload;outeractivelement:body;starttest;middleactivelement:body;callfocus;middleactivelement:iframe;innerfocus;middletimer;middleactivelement:iframe;outeractivelement:iframe;" but got "outeronload;outeractivelement:body;starttest;middleactivelement:body;callfocus;middleactivelement:iframe;middletimer;middleactivelement:iframe;outeractivelement:iframe;"
Harness: the test ran to completion.
